Move models to new shared folder
