+++ projects/fhi-angular-components/src/lib/fhi-time/fhi-week-range/fhi-week-range.model.ts
@@ -1,4 +1,4 @@
-import { FhiWeek } from '../fhi-week.model';
+import { FhiWeek } from '../shared/models/fhi-week.model';
 
 export class FhiWeekRange {
   from: FhiWeek;
