+++ projects/fhi-angular-components/src/lib/fhi-time/fhi-year-month-range/fhi-month-range.model.ts
@@ -1,4 +1,4 @@
-import { FhiMonth } from '../fhi-month.model';
+import { FhiMonth } from '../shared/models/fhi-month.model';
 
 export class FhiMonthRange {
   from: FhiMonth;
